#172: Rewrote both getAllValues() methods on ByName and ByToken to an implementation with Predicate<ParseValue>.
