+++ core/src/main/java/io/parsingdata/metal/data/selection/ByName.java
@@ -16,12 +16,9 @@
 
 package io.parsingdata.metal.data.selection;
 
-import static io.parsingdata.metal.SafeTrampoline.complete;
-import static io.parsingdata.metal.SafeTrampoline.intermediate;
 import static io.parsingdata.metal.Util.checkNotNull;
 import static io.parsingdata.metal.data.transformation.Reversal.reverse;
 
-import io.parsingdata.metal.SafeTrampoline;
 import io.parsingdata.metal.data.ImmutableList;
 import io.parsingdata.metal.data.ParseGraph;
 import io.parsingdata.metal.data.ParseItem;
@@ -60,26 +57,7 @@ public final class ByName {
     public static ImmutableList<Value> getAllValues(final ParseGraph graph, final String name) {
         checkNotNull(graph, "graph");
         checkNotNull(name, "name");
-        return reverse(getAllValuesRecursive(ImmutableList.create(graph), new ImmutableList<>(), name).computeResult());
-    }
-
-    private static SafeTrampoline<ImmutableList<Value>> getAllValuesRecursive(final ImmutableList<ParseGraph> graphList, final ImmutableList<Value> valueList, final String name) {
-        if (graphList.isEmpty()) { return complete(() -> valueList); }
-        final ParseGraph graph = graphList.head;
-        if (graph.isEmpty()) { return intermediate(() -> getAllValuesRecursive(graphList.tail, valueList, name)); }
-        return intermediate(() -> getAllValuesRecursive(addIfGraph(graphList.tail.add(graph.tail), graph.head),
-                                                        addIfMatchingValue(valueList, graph.head, name),
-                                                        name));
-    }
-
-    private static ImmutableList<ParseGraph> addIfGraph(final ImmutableList<ParseGraph> graphList, final ParseItem item) {
-        if (item.isGraph()) { return graphList.add(item.asGraph()); }
-        return graphList;
-    }
-
-    private static ImmutableList<Value> addIfMatchingValue(final ImmutableList<Value> valueList, final ParseItem item, final String name) {
-        if (item.isValue() && item.asValue().matches(name)) { return valueList.add(item.asValue()); }
-        return valueList;
+        return reverse(ByPredicate.getAllValues(ImmutableList.create(graph), new ImmutableList<>(), (value) -> value.matches(name)).computeResult());
     }
 
     public static ParseValue get(final ImmutableList<ParseValue> list, final String name) {
